+++ src/main/java/sa/gov/nic/bio/bw/client/features/faceverification/MatchingFxController.java
@@ -86,7 +86,7 @@ public class MatchingFxController extends WizardStepFxControllerBase
 		
 		Map<String, Object> uiDataMap = new HashMap<>();
 		uiDataMap.put(Workflow.KEY_WEBSERVICE_RESPONSE, null);
-		Context.getWorkflowManager().submitUserTask(uiDataMap);
+		if(!isDetached()) Context.getWorkflowManager().submitUserTask(uiDataMap);
 	}
 	
 	private void showProgress(boolean bShow)