+++ src/main/java/sa/gov/nic/bio/bw/client/features/fingerprintcardidentification/InquiryResultPaneFxController.java
@@ -28,8 +28,7 @@ import sa.gov.nic.bio.bw.client.features.commons.webservice.Name;
 import sa.gov.nic.bio.bw.client.features.commons.webservice.PersonIdInfo;
 import sa.gov.nic.bio.bw.client.features.commons.webservice.PersonInfo;
 import sa.gov.nic.bio.bw.client.features.fingerprintcardidentification.beans.FingerprintCardIdentificationRecordReport;
-import sa.gov.nic.bio.bw.client.features.fingerprintcardidentification.tasks.
-																	BuildFingerprintCardIdentificationRecordReportTask;
+import sa.gov.nic.bio.bw.client.features.fingerprintcardidentification.tasks.BuildFingerprintCardIdentificationRecordReportTask;
 import sa.gov.nic.bio.bw.client.features.fingerprintcardidentification.utils.FingerprintCardIdentificationErrorCodes;
 
 import java.io.ByteArrayInputStream;
@@ -42,7 +41,6 @@ import java.util.Base64;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
-import java.util.Locale;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicReference;
 
@@ -283,28 +281,29 @@ public class InquiryResultPaneFxController extends WizardStepFxControllerBase
 			if(name != null)
 			{
 				String sFirstName = name.getFirstName();
-				if(sFirstName != null && !sFirstName.trim().isEmpty())
+				if(sFirstName != null && !sFirstName.trim().isEmpty() && !sFirstName.trim().equals("-"))
 				{
 					firstName = sFirstName;
 					lblFirstName.setText(firstName);
 				}
 				
 				String sFatherName = name.getFatherName();
-				if(sFatherName != null && !sFatherName.trim().isEmpty())
+				if(sFatherName != null && !sFatherName.trim().isEmpty() && !sFatherName.trim().equals("-"))
 				{
 					fatherName = sFatherName;
 					lblFatherName.setText(fatherName);
 				}
 				
 				String sGrandfatherName = name.getGrandfatherName();
-				if(sGrandfatherName != null && !sGrandfatherName.trim().isEmpty())
+				if(sGrandfatherName != null && !sGrandfatherName.trim().isEmpty() &&
+																				!sGrandfatherName.trim().equals("-"))
 				{
 					grandfatherName = sGrandfatherName;
 					lblGrandfatherName.setText(grandfatherName);
 				}
 				
 				String sFamilyName = name.getFamilyName();
-				if(sFamilyName != null && !sFamilyName.trim().isEmpty())
+				if(sFamilyName != null && !sFamilyName.trim().isEmpty() && !sFamilyName.trim().equals("-"))
 				{
 					familyName = sFamilyName;
 					lblFamilyName.setText(familyName);
@@ -346,14 +345,14 @@ public class InquiryResultPaneFxController extends WizardStepFxControllerBase
 				String sOccupation = identityInfo.getOccupation();
 				if(sOccupation != null && !sOccupation.trim().isEmpty())
 				{
-					occupation = sOccupation;
+					occupation = AppUtils.localizeNumbers(sOccupation);
 					lblOccupation.setText(occupation);
 				}
 				
 				String sIdNumber = identityInfo.getIdNumber();
 				if(sIdNumber != null && !sIdNumber.trim().isEmpty())
 				{
-					idNumber = sIdNumber;
+					idNumber = AppUtils.localizeNumbers(sIdNumber);
 					lblIdNumber.setText(idNumber);
 				}
 				
@@ -376,7 +375,7 @@ public class InquiryResultPaneFxController extends WizardStepFxControllerBase
 					
 					if(theIdType != null)
 					{
-						idType = AppUtils.replaceNumbersOnly(theIdType.getDesc(), Locale.getDefault());
+						idType = AppUtils.localizeNumbers(theIdType.getDesc());
 						lblIdType.setText(idType);
 					}
 				}
@@ -403,7 +402,7 @@ public class InquiryResultPaneFxController extends WizardStepFxControllerBase
 			String sBirthPlace = personInfo.getBirthPlace();
 			if(sBirthPlace != null && !sBirthPlace.trim().isEmpty())
 			{
-				birthPlace = sBirthPlace;
+				birthPlace = AppUtils.localizeNumbers(sBirthPlace);
 				lblBirthPlace.setText(birthPlace);
 			}
 			
@@ -418,7 +417,7 @@ public class InquiryResultPaneFxController extends WizardStepFxControllerBase
 
 		if(idNumber == null)
 		{
-			idNumber = AppUtils.replaceNumbersOnly(String.valueOf(samisId), Locale.getDefault());
+			idNumber = AppUtils.localizeNumbers(String.valueOf(samisId));
 			idType = null;
 			lblIdNumber.setText(idNumber);
 			lblIdType.setText(idType);